backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
tavily-python
pymongo
python-dotenv
beautifulsoup4
pydantic>=2.12,<3
langchain-tavily

